refactor(v4): split the compiler's emitters per core subclass in core/jit.ts
Revives eae81430 (branch tuple-jit) on top of main, as a table rather than as constructor twins. Every core class gets its own fast-path emitter in core/jit.ts, keyed by the trait name its constructor records; `codegenFor` resolves a schema's emitter from its own `_zod.codegen` override or the most specific recorded trait, so `generateCheck` no longer switches on def.type. Classic and mini keep constructing the plain core classes: constructing twins retained the whole compiler in every classic bundle (+5.7 KB gz on a bundle that never calls z.compile) and left mini's z.compile compiling nothing, which is what the revert bf7f5632 avoided. Bundles without z.compile are byte-identical to main; with it, +340 B gz for the table.

Carries main's changes to the moved generators (addUserConstant sites, intersection and record `definite` carry-out, the properties schema) and adds a `$ZodProperties` emitter.
